requirements.txt
Django>=1.10,<1.11
waitress>=0.8,<0.9
whitenoise>=3.2,<3.3

dj-database-url==0.4.1
django-debug-toolbar>=1.6,<1.7

python-decouple==3.0
pytz==2015.7
sqlparse>=0.2,<0.3

psycopg2>=2.6,<2.7

redis>=2.10,<2.11
django-redis>=4.3,<4.4

rq>=0.5,<0.6
rq-scheduler>=0.6,<0.7
django-rq>=0.9,<0.10

raven>=5.16,<5.17

django-import-export>=0.4,<0.5

opengraph>=0.5,<0.6

facebook-sdk>=2.0,<2.1

crontab>=0.21,<0.22

requests>=2.11,<2.12

pytest-django>=3.0,<3.1
pytest-httpretty>=0.2,<0.3
factory_boy>=2.7,<2.8
